requirements.txt
torch
torchvision
torchaudio
hydra-core
hydra-submitit-launcher
mlflow
dagshub
tqdm
